% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_name_abbr.R
\name{abbr2name}
\alias{abbr2name}
\title{Abbreviation to full state name}
\usage{
abbr2name(abbr)
}
\arguments{
\item{abbr}{Abbreviation}
}
\value{
Full state name
}
\description{
Abbreviation to full state name
}
\examples{
abbr2name("CA")
}
