% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_wage.R
\name{find_difs}
\alias{find_difs}
\title{Find difference of minimum wage}
\usage{
find_difs(
  start_year,
  end_year,
  data = minimum_wage,
  kind = "Effective.Minimum.Wage"
)
}
\arguments{
\item{start_year}{The start year}

\item{end_year}{The end year}

\item{data}{Minimum wage}
}
\value{
The difference of minimum wage
}
\description{
Find difference of minimum wage.
}
\examples{
find_difs(2010, 2015)
}
