% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_code.R
\name{fips2county}
\alias{fips2county}
\title{Turn FIPS Code to County Name}
\usage{
fips2county(fips)
}
\arguments{
\item{fips}{FIPS codes are numbers which uniquely identify geographic areas.}
}
\value{
The county name of the county
}
\description{
Return the county name of a county according to the given FIPS code.
}
\examples{
fips2county('01001')
}
