% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_code.R
\name{fips2name}
\alias{fips2name}
\title{Turn FIPS Code to Full Name}
\usage{
fips2name(fips)
}
\arguments{
\item{fips}{FIPS codes are numbers which uniquely identify geographic areas.}
}
\value{
The full name of the county
}
\description{
Return the name of a county according to the given FIPS code.
}
\examples{
fips2name('01001')
}
