% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_code.R
\name{fips2region}
\alias{fips2region}
\title{Turn FIPS Code to Region Name}
\usage{
fips2region(fips)
}
\arguments{
\item{fips}{FIPS codes are numbers which uniquely identify geographic areas.}
}
\value{
The region name of the county
}
\description{
Return the region name of a county according to the given FIPS code.
}
\examples{
fips2region('01001')
}
