% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_code.R
\name{fips2state}
\alias{fips2state}
\title{Turn FIPS Code to State Name}
\usage{
fips2state(fips)
}
\arguments{
\item{fips}{FIPS codes are numbers which uniquely identify geographic areas.}
}
\value{
The state name of the county
}
\description{
Return the state name of a county according to the given FIPS code.
}
\examples{
fips2state('01001')
}
