% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_code.R
\docType{data}
\name{fips_code}
\alias{fips_code}
\title{The FIPS Code of U.S.}
\format{
A data frame with 3145 rows and 13 variables.
}
\source{
<https://github.com/kjhealy/fips-codes>
}
\usage{
fips_code
}
\description{
A data set with FIPS Code and Geographical information.
}
\keyword{datasets}
