% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs5_data.R
\name{get_acs5_data}
\alias{get_acs5_data}
\title{Get ACS Data}
\usage{
get_acs5_data(table_name, period, data_code)
}
\arguments{
\item{table_name}{The ACS table name of the dataset}

\item{period}{The time period of the dataset}

\item{data_code}{The ACS Code of the variable}
}
\value{
ACS Data
}
\description{
Get American Community Survey (ACS) Data from censusapi.
}
\examples{
get_acs5_data("acs/acs5", 2010:2015, c("B20017A", "B20017B"))
}
