% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs5_data.R
\name{get_acs5_distrubution}
\alias{get_acs5_distrubution}
\title{Get ACS Distrubution Data}
\usage{
get_acs5_distrubution(table_name, period, data_code)
}
\arguments{
\item{table_name}{The ACS table name of the dataset}

\item{period}{The time period of the dataset}

\item{data_code}{The ACS Code of the variable}
}
\value{
ACS Data
}
\description{
Get ACS Distrubution Data
}
