% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bea_data.R
\name{get_bea_data}
\alias{get_bea_data}
\title{Get Bea Data}
\usage{
get_bea_data(beaKey, table_name, line_code, period)
}
\arguments{
\item{beaKey}{API key}

\item{table_name}{The table name of the data}

\item{line_code}{The line of the data}

\item{period}{The time period of the data}
}
\value{
A data set from BEA
}
\description{
Get data from U.S. Bureau of Economic Analysis (BEA) with API
}
\examples{
get_bea_data("abcdefghijklmn", "CAGDP1", 1, 2010:2015)
}
