% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unemp_from_bls.R
\name{get_unemp_from_bls}
\alias{get_unemp_from_bls}
\title{Get unemployment data}
\usage{
get_unemp_from_bls(urls)
}
\arguments{
\item{urls}{The list of urls <example.xlsx> where unemployment data are stored.}
}
\value{
The unemployment dataset
}
\description{
Get unemployment data from U.S. Bureau of Labor Statistics website
}
\examples{
get_unemp_from_bls("https://www.bls.gov/lau/laucnty20.xlsx")
}
