% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landarea.R
\docType{data}
\name{landarea}
\alias{landarea}
\title{Landarea by County}
\format{
A data frame with 3198 rows and 3 variables.
}
\source{
<https://www2.census.gov/library/publications/2011/compendia/usa-counties/excel/LND01.xls>
}
\usage{
landarea
}
\description{
A data set with the FIPS code and the area of the county in sq mi.
}
\keyword{datasets}
