% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_wage.R
\docType{data}
\name{minimum_wage}
\alias{minimum_wage}
\title{The Minumum wage by year and state}
\format{
A data frame with 2862 rows and 7 variables.
}
\source{
<https://www.dol.gov/agencies/whd/state/minimum-wage/history>
}
\usage{
minimum_wage
}
\description{
A data set containing historical federal minimum wage and it of each state.
}
\keyword{datasets}
