% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_name_abbr.R
\name{name2abbr}
\alias{name2abbr}
\title{State name to Abbreviation}
\usage{
name2abbr(name)
}
\arguments{
\item{name}{State name}
}
\value{
Abbreviation
}
\description{
State name to Abbreviation
}
\examples{
abbr2name("California")
}
