% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotencoding.R
\name{onehotencoding}
\alias{onehotencoding}
\title{One hot encoding}
\usage{
onehotencoding(dataset, column)
}
\arguments{
\item{dataset}{A data set}

\item{column}{A one hot encoded column}
}
\value{
A data set with one hot encoded variables
}
\description{
One hot encoding
}
\examples{
onehotencoding(mtcars, "gear")
}
