% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_name_abbr.R
\name{state2area}
\alias{state2area}
\title{Find area of a state}
\usage{
state2area(state)
}
\arguments{
\item{state}{State name}
}
\value{
The area in sq mi of the state
}
\description{
Find area of a state
}
\examples{
state2area("California")
}
